cleanup db platform-specific traits
