+++ xmtp_db/src/encrypted_store/database/native/sqlcipher_connection.rs
@@ -12,9 +12,10 @@ use std::{
     path::{Path, PathBuf},
 };
 
-use crate::{NotFound, native::NativeStorageError};
+use super::NativeStorageError;
+use crate::NotFound;
 
-use super::{EncryptionKey, StorageOption};
+use crate::{EncryptionKey, StorageOption};
 
 pub type Salt = [u8; 16];
 const PLAINTEXT_HEADER_SIZE: usize = 32;
@@ -45,7 +46,7 @@ pub struct EncryptedConnection {
 impl EncryptedConnection {
     /// Creates a file for the salt and stores it
     pub fn new(key: EncryptionKey, opts: &StorageOption) -> Result<Self, NativeStorageError> {
-        use super::StorageOption::*;
+        use crate::StorageOption::*;
 
         let salt = match opts {
             Ephemeral => None,
@@ -256,7 +257,7 @@ impl EncryptedConnection {
     }
 }
 
-impl super::native::ValidatedConnection for EncryptedConnection {
+impl super::ValidatedConnection for EncryptedConnection {
     fn validate(&self, opts: &StorageOption) -> Result<(), NativeStorageError> {
         let conn = &mut opts.conn()?;
         let sqlcipher_version = EncryptedConnection::check_for_sqlcipher(opts, Some(conn))?;
