+++ xmtp_db/src/encrypted_store/database/wasm.rs
@@ -1,6 +1,6 @@
 //! WebAssembly specific connection for a SQLite Database
 //! Stores a single connection behind a mutex that's used for every libxmtp operation
-use super::{StorageOption, XmtpDb, db_connection::DbConnectionPrivate};
+use crate::{StorageOption, XmtpDb, db_connection::DbConnectionPrivate};
 use diesel::prelude::SqliteConnection;
 use diesel::{connection::AnsiTransactionManager, prelude::*};
 use parking_lot::Mutex;
@@ -106,7 +106,7 @@ impl std::fmt::Debug for WasmDb {
 
 impl WasmDb {
     pub async fn new(opts: &StorageOption) -> Result<Self, WasmStorageError> {
-        use super::StorageOption::*;
+        use crate::StorageOption::*;
         init_sqlite().await;
         maybe_resize().await?;
         let conn = match opts {
