+++ xmtp_db/src/encrypted_store/db_connection.rs
@@ -12,7 +12,7 @@ use std::{
 use super::XmtpDb;
 
 #[cfg(not(target_arch = "wasm32"))]
-pub type DbConnection = DbConnectionPrivate<super::RawDbConnection>;
+pub type DbConnection = DbConnectionPrivate<crate::database::RawDbConnection>;
 
 #[cfg(target_arch = "wasm32")]
 pub type DbConnection = DbConnectionPrivate<diesel::prelude::SqliteConnection>;
